GUI updates: make things work more or less
Will rewrite parts of this logic though

+++ core/BDMORPH.h
@@ -100,7 +100,7 @@ public:
 	BDMORPHModel(MeshModel& orig);
 	~BDMORPHModel();
 
-	int interpolate_frame(MeshModel *a, MeshModel* b, double t);
+	double interpolate_frame(MeshModel *a, MeshModel* b, double t);
 private:
 
 	OrderedEdge e0;
